back/requirements.txt
Django
djangorestframework
django-cors-headers
pymysql
pycryptodome
PyJWT
djangorestframework-simplejwt